Moved inner components out of Shell
