+++ client/src/components/App/App.js
@@ -3,15 +3,14 @@ import React from 'react';
 import { withStyles } from '@material-ui/core';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 import NotFound from '../statuscodes/404/NotFound';
-import Dashboard from '../Shell/CustomerPortal/MenuOptions/Dashboard';
-import Account from '../Shell/CustomerPortal/MenuOptions/Account';
-import RegStudForm from '../Shell/CustomerPortal/MenuOptions/RegisteredStudentsForm/RegStudForm';
-import FAQ from '../Shell/CustomerPortal/MenuOptions/FAQ';
-import AboutUs from '../Shell/CustomerPortal/MenuOptions/AboutUs';
-import EventsCalendar from '../Shell/CustomerPortal/MenuOptions/EventsCalendar';
-import EventRegistration from '../Shell/CustomerPortal/MenuOptions/EventRegistration';
-import ActiveRegistration from '../Shell/CustomerPortal/MenuOptions/ActiveRegistration/ActiveRegistrations';
-// import CustomerPortalContainer from '../../containers/Shell/CustomerPortal/CustomerPortalContainer';
+import Dashboard from '../CustomerPortal/Dashboard/Dashboard';
+import Account from '../CustomerPortal/Account/Account';
+import RegStudForm from '../CustomerPortal/RegisteredStudentsForm/RegStudForm';
+import FAQ from '../CustomerPortal/FAQ/FAQ';
+import AboutUs from '../CustomerPortal/AboutUs/AboutUs';
+import EventsCalendar from '../CustomerPortal/EventsCalendar/EventsCalendar';
+import EventRegistration from '../CustomerPortal/EventRegistration/EventRegistration';
+import ActiveRegistration from '../CustomerPortal/ActiveRegistration/ActiveRegistrations';
 import Login from '../Login/Login';
 
 const styles = () => ({
