+++ client/src/components/CustomerPortal/AboutUs/AboutUs.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { withStyles } from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
-import CustomerPortalContainer from '../../../../containers/Shell/CustomerPortal/CustomerPortalContainer';
+import CustomerPortalContainer from '../../../containers/Shell/CustomerPortal/CustomerPortalContainer'; 
 
 const styles = theme => ({
   panel: {
