--- client/src/components/Shell/CustomerPortal/tiledata.js
@@ -1,70 +1,0 @@
-// import React from 'react';
-// import { Link } from 'react-router-dom';
-// import { Icon } from '@material-ui/core';
-// import ListItem from '@material-ui/core/ListItem';
-// import ListItemIcon from '@material-ui/core/ListItemIcon';
-// import ListItemText from '@material-ui/core/ListItemText';
-
-// export const mailFolderListItems = ({ props }) => (
-//   <div>
-//     <ListItem button component={Link} to="/customer">
-//       <ListItemIcon>
-//         <Icon>dashboard</Icon>
-//       </ListItemIcon>
-//       <ListItemText primary="Dashboard" />
-//     </ListItem>
-//     <ListItem button component={Link} to="/customer/account-info">
-//       <ListItemIcon>
-//         <Icon>account_box</Icon>
-//       </ListItemIcon>
-//       <ListItemText primary="Your Account" />
-//     </ListItem>
-//     <ListItem button component={Link} to="/customer/registered-students">
-//       <ListItemIcon>
-//         <Icon>group_add</Icon>
-//       </ListItemIcon>
-//       <ListItemText primary="Registered Students" />
-//     </ListItem>
-//     <ListItem button component={Link} to="/customer/events-calendar">
-//       <ListItemIcon>
-//         <Icon>calendar_today</Icon>
-//       </ListItemIcon>
-//       <ListItemText primary="Events Calendar" />
-//     </ListItem>
-//     <ListItem button component={Link} to="/customer/register-for-an-event">
-//       <ListItemIcon>
-//         <Icon>event_available</Icon>
-//       </ListItemIcon>
-//       <ListItemText primary="Register for an Event" />
-//     </ListItem>
-//     <ListItem button component={Link} to="/customer/active-registrations">
-//       <ListItemIcon>
-//         <Icon>event</Icon>
-//       </ListItemIcon>
-//       <ListItemText primary="Active Registrations" />
-//     </ListItem>
-//     <ListItem button component={Link} to="/customer/payment-history">
-//       <ListItemIcon>
-//         <Icon>payment</Icon>
-//       </ListItemIcon>
-//       <ListItemText primary="Payment History" />
-//     </ListItem>
-//   </div>
-// );
-
-// export const helperItems = (
-//     <div>
-//     <ListItem button component={Link} to="/customer/faq">
-//       <ListItemIcon>
-//         <Icon>help_outline</Icon>
-//       </ListItemIcon>
-//       <ListItemText primary="Frequently Asked Questions" />
-//     </ListItem>
-//     <ListItem button component={Link} to="/customer/about">
-//       <ListItemIcon>
-//         <Icon>music_note</Icon>
-//       </ListItemIcon>
-//       <ListItemText primary="About Us" />
-//     </ListItem>
-//     </div>
-// );
