+++ client/src/containers/Shell/CustomerPortal/CustomerPortalContainer.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 import { togglePortalDrawer } from '../../../actions/CustomerPortalActions';
 
-import CustomerDrawer from '../../../components/Shell/CustomerPortal/CustomerPortal';
+import CustomerDrawer from '../../../components/CustomerPortal/CustomerPortal';
 
 const mapStateToProps = ({ CustomerPortal: { open } }) => ({ open });
